OTP Update
I have built out the OTP form and functions needed to validate a users OTP activation process. The changes included are:

1) Created OTP files x 4
2) Updated the app.module to include the new files in the component declarations
3) OTP html includes standard detail - two fields for email and OTP
4) CSS detail
5) OTP module file
6) ts file with appropriate methods
7) updated the auth service with the post
8) Updated the signup ts file to direct a new user registering to the OTP activation screen submiting registration records

Tested - all test cases work well validating if email exists, if otp record is correct, if the user is already active. the outcome of a successful action is the user record is updated to active in the DB - success.